Split MinioBackendRegistry tests into separate unit and integration tests
