+++ mindtrace/core/mindtrace/core/config/config.py
@@ -6,6 +6,9 @@ class Config(dict):
             "MINDTRACE_TEMP_DIR": "~/.cache/mindtrace/temp",
             "MINDTRACE_DEFAULT_REGISTRY_DIR": "~/.cache/mindtrace/registry",
             "MINDTRACE_MINIO_REGISTRY_URI": "~/.cache/mindtrace/minio-registry",
+            "MINDTRACE_MINIO_ENDPOINT": "localhost:9000",
+            "MINDTRACE_MINIO_ACCESS_KEY": "minioadmin",
+            "MINDTRACE_MINIO_SECRET_KEY": "minioadmin",
         }
         # Update defaults with any provided kwargs
         default_config.update(kwargs)
